+++ cli/src/fmt/simple/mod.rs
@@ -2,4 +2,5 @@ pub struct Simple;
 
 mod account;
 mod asset;
+mod trade_aggregation;
 mod transaction;